separated header options buttons. linked marshall pic to profile-page
