+++ src/components/Header/Actions.jsx
@@ -1,5 +1,6 @@
 import NotifDropdownMenu from "./NotifDropdownMenu.jsx";
 import "./Actions.css";
+import { Link } from "react-router-dom";
 
 export default function Actions() {
   return (
@@ -10,10 +11,12 @@ export default function Actions() {
       </button>
       <div className="account">
         <span className="greeting">Hello, Marshal</span>
+        <Link to="/profile-page">
         <img
           src="https://i.pinimg.com/736x/63/d4/2b/63d42bcfc3ce97414d78f14ae76f61c8.jpg"
           alt="Profile Picture"
         />
+        </Link>
         <div className="account-panel">
           <div className="account-info">
             <img
