+++ src/components/ProfilePage/HeaderOptions.css
@@ -1,0 +1,43 @@
+.headerOptions{
+	margin-top: 27px;
+	size: 5px;
+	gap: 3px;
+	padding-left: 5px;
+	display: block;
+}
+
+.headerOptions button{
+	display: inline-block;
+	margin: 0px 3px;
+	padding: 0px .5rem;
+	height: 25px;
+	background-color: var(--primary-color);
+	border-radius: 15px;
+	border-spacing: .5px;	
+	text-align: right;
+}
+.headerOptions button:hover{
+	transition: background-color 250ms;
+	background: var(--secondary-color);
+	
+}
+
+.headerOptions-edit{
+	margin-top: 27px;
+	margin-left: 50px;
+	size: 5px;
+	gap: 3px;
+	padding-left: 5px;
+	display: block;
+}
+
+.headerOptions-edit button{
+	display: inline-block;
+	margin: 0px 3px;
+	padding: 0px .5rem;
+	height: 25px;
+	background-color: var(--primary-color);
+	border-radius: 15px;
+	border-spacing: .5px;	
+	text-align: right;
+}
