+++ src/components/ProfilePage/ProfileHeader.css
@@ -51,27 +51,5 @@
 	position: relative;
 }
 
-.headerOptions{
-	margin-top: 50px;
-	size: 5px;
-	gap: 3px;
-	padding-left: 5px;
-	display: block;
-}
 
-.headerOptions button{
-	display: inline-block;
-	margin: 0px 3px;
-	padding: 0px .5rem;
-	height: 25px;
-	background-color: var(--primary-color);
-	border-radius: 15px;
-	border-spacing: .5px;	
-	text-align: right;
-}
-.headerOptions button:hover{
-	transition: background-color 250ms;
-	background: var(--secondary-color);
-	
-}
