+++ src/components/ProfilePage/ProfileHeader.jsx
@@ -1,23 +1,11 @@
 import "./ProfileHeader.css";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import {
-	faEllipsisH,
-	faCommenting,
-	faPlus,
-	faCheck
-} from "@fortawesome/free-solid-svg-icons";
-import { useState } from "react";
+import HeaderOptions from "./HeaderOptions";
 
 
 export default function ProfileHeader() {
-	
-	const [followBtn, setFollowBtn]=useState(false);
-	function handleClick(){
-		setFollowBtn(followBtn => !followBtn);
-	};
+	const headerOptions = HeaderOptions(false);	
 
-	let followBtnCheck = followBtn ? '-follow' : '-following';
-		return (
+	return (
 	  <div className="profileHeader">
 		<div className="coverPhoto"></div>
 		<div className="info-container">
@@ -28,11 +16,7 @@ export default function ProfileHeader() {
 					<div className="bio">2nd Year BSCS-ST at DLSU <br/>dm for Math Tutoring</div>
 				</div>
 			<div className="headerOptions">
-				<button><FontAwesomeIcon icon={faEllipsisH} id="elipsisIcon"></FontAwesomeIcon></button>
-				<button><FontAwesomeIcon icon={faCommenting} id="message"></FontAwesomeIcon></button>
-				<button onClick={handleClick} className={`btn${followBtnCheck}`}>
-					<span className="followState"> {followBtn ? (<span><FontAwesomeIcon icon={faPlus} id="plusIcon"></FontAwesomeIcon> Follow</span>) : (<span><FontAwesomeIcon icon={faCheck} id="checkIcon"></FontAwesomeIcon> Following</span>)}</span>
-					</button>
+				{headerOptions}
 				</div>
 			</div>
